db: change types to avoid headers changing between
configure wants to use host-specific types to get a 64-bit integer in db.h
instead of using an alias such as int64_t.  This means that the header differs
in multilib environments for no good reason, so replace the type with the alias
in stdint.h.

This then breaks the overly complicated type check but as we know that int64_t
exists and works, we can just delete that.

Upstream-Status: Pending
Signed-off-by: Ross Burton <ross.burton@intel.com>

%% original patch: sequence-type.patch